+++ src/app/category/[slug]/CategoryPage.module.css
@@ -1,4 +1,4 @@
-/* styles/CategoryPage.module.css */
+/* src/app/category/[slug]/CategoryPage.module.css */
 .pageContainer {
   /* Optional: Add overall page padding or max-width if desired */
 }
