+++ src/components/FeaturedCategories.module.css
@@ -1,4 +1,4 @@
-/* styles/FeaturedCategories.module.css */
+/* src/components/FeaturedCategories.module.css */
 .featuredCategories {
   padding: 20px;
   background-color: #f9f9f9; /* Light mode background */
