+++ src/components/FeaturedProductsCarousel.module.css
@@ -1,4 +1,4 @@
-/* styles/FeaturedProductsCarousel.module.css */
+/* src/components/FeaturedProductsCarousel.module.css */
 .featuredProductsCarousel {
   margin-bottom: 2rem;
 }
