+++ src/components/Navbar.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import { useState } from 'react';
-import styles from '../../styles/Navbar.module.css';
+import styles from './Navbar.module.css';
 import { Category } from '@/types'; // Import Category type
 
 interface NavbarProps {
